Extract status related stuff into its own helper
